Added @media query: landscape orientation.
Added @media query for landscape orientation.  Previous version's parallax layer was too short when in landscape orientation, as it was set to 100% of height (now set to 210% of height).  100% of height is proper in typical portraint orientation, given the tall & narrow portrait proportions; the wide & stunted proportions of landscape orientation at 100% height was too short for the html content.